@@ -109,26 +110,31 @@ dependencies:
   wifi_scan: ^0.4.1
 
   #  media
-  youtube_player_flutter: ^8.1.2
-  youtube_explode_dart: ^1.12.4
-  unique_identifier: ^0.3.0
+  youtube_player_flutter: ^9.1.1
+  youtube_explode_dart: ^2.4.0
+  unique_identifier: 0.4.0
   collection: any
-  disk_space: ^0.2.1
+#  disk_space_plus: 0.2.4
+  disk_space_2: ^1.0.8
 
+  screen_control:
+    git:
+      url: https://github.com/mawaqit/native_screen_lock_mawaqit_devices
+      ref: master
   equatable: ^2.0.5
-  just_audio: ^0.9.37
+  just_audio: ^0.10.0
 
 
   # alert
   flash: ^3.0.5+2
 
   # app update
-  upgrader: ^6.5.0
+  upgrader: ^11.3.1
   open_store: ^0.5.0
 
   # internet connectivity checker
-  internet_connection_checker_plus: ^1.0.1
-  flutter_kurdish_localization: ^1.0.8
+  internet_connection_checker_plus: ^2.7.1
+  #  flutter_kurdish_localization: ^1.0.8
   scrollview_observer: ^1.21.0
   # manage files and folders
   flutter_archive: ^6.0.3
