@@ -1,5 +1,6 @@
 import 'dart:developer';
 import 'package:flutter/material.dart';
+import 'package:flutter_background_service/flutter_background_service.dart';
 import 'package:global_configuration/global_configuration.dart';
 import 'package:mawaqit/i18n/l10n.dart';
 import 'package:mawaqit/src/const/config.dart';
@@ -7,6 +8,7 @@ import 'package:mawaqit/src/const/constants.dart';
 import 'package:mawaqit/src/helpers/AnalyticsWrapper.dart';
 import 'package:provider/provider.dart';
 import 'package:shared_preferences/shared_preferences.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 import '../src/services/mosque_manager.dart';
 
@@ -74,7 +76,11 @@ class AppLanguage extends ChangeNotifier {
       language: languageName(type.languageCode),
       mosqueId: mosqueId,
     );
-
+    final service = FlutterBackgroundService();
+    service.invoke(
+      'updateLocalization',
+      {'language_code': type.languageCode},
+    );
     if (S.supportedLocales.indexOf(type) != -1) {
       _appLocale = type;
       await prefs.setString('language_code', type.languageCode);
