@@ -41,21 +41,19 @@ abstract class RandomHadithConstant {
 }
 
 class TurnOnOffTvConstant {
-  static const String kLastEventDate = "lastEventDate";
-  static const String kIsEventsSet = "isEventsSet";
-  static const String kActivateToggleFeature = "activateToggleFeature";
-  static const String kScheduledTimersKey = "scheduledTimers";
+  static const String kActivateToggleFeature = 'activate_toggle_feature';
+  static const String kisFajrIshaOnly = 'is_fajr_isha_only';
+  static const String kScheduledTimersKey = 'scheduled_timers_key';
+  static const String kLastEventDate = 'last_event_date';
+  static const String kMinuteBeforeKey = 'minute_before_key';
+  static const String kMinuteAfterKey = 'minute_after_key';
+  static const String kIsEventsSet = 'is_events_set';
+  static const String kScheduleParamsKey = 'schedule_params_key';
+  static const String kLastExecutedEventDate = 'last_executed_event_date';
 
   /// native methods calls
   static const String kNativeMethodsChannel = "nativeMethodsChannel";
   static const String kCheckRoot = "checkRoot";
-  static const String kToggleBoxScreenOff = "toggleBoxScreenOff";
-  static const String kToggleBoxScreenOn = "toggleBoxScreenOn";
-  static const String kToggleTabletScreenOn = "toggleTabletScreenOn";
-  static const String kToggleTabletScreenOff = "toggleTabletScreenOff";
-  static const String kisFajrIshaOnly = "isIshaFajrOnly";
-  static const String kMinuteBeforeKey = 'selectedMinuteBefore';
-  static const String kMinuteAfterKey = 'selectedMinuteAfter';
 }
 
 abstract class AnnouncementConstant {
@@ -81,6 +79,8 @@ abstract class QuranConstant {
   static const String quranMoshafConfigJsonUrl = 'https://cdn.mawaqit.net/quran/tv_config.json';
   static const String kIsFirstTime = 'is_first_time_quran';
   static const String kQuranReciterImagesBaseUrl = 'https://cdn.mawaqit.net/quran/reciters-pictures/';
+  static const String kQuranCacheBoxName = 'timestamp_box';
+  static const String kQuranReciterRetentionTime = 'quran_reciter_retention_time';
   static const int kCacheWidth = 50;
   static const int kCacheHeight = 50;
 }
