@@ -1,12 +1,16 @@
+import 'dart:developer';
 import 'dart:io';
 
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:fpdart/fpdart.dart';
 import 'package:mawaqit/src/data/data_source/quran/recite_remote_data_source.dart';
 import 'package:mawaqit/src/data/data_source/quran/reciter_local_data_source.dart';
+import 'package:mawaqit/src/domain/error/recite_exception.dart';
 
 import 'package:mawaqit/src/domain/model/quran/reciter_model.dart';
 
 import 'package:mawaqit/src/domain/repository/quran/recite_repository.dart';
+import 'package:mawaqit/src/helpers/AppDate.dart';
 
 import '../../../domain/model/quran/audio_file_model.dart';
 
@@ -19,14 +23,50 @@ class ReciteImpl implements ReciteRepository {
   @override
   Future<List<ReciterModel>> getAllReciters({required String language}) async {
     try {
+      // Check if reciters are cached locally
+      final cachedReciters = await _localDataSource.getReciters();
+      final lastUpdatedOption = _localDataSource.getLastUpdatedTimestamp();
+
+      // Check if the cached data is still valid (within 1 month)
+      final isCacheValid = lastUpdatedOption.match(
+        () => false, // No timestamp available, cache is invalid
+        (lastUpdated) {
+          final oneMonthAgo = DateTime.now().subtract(Duration(days: 30)); // 1 month ago
+          return lastUpdated.isAfter(oneMonthAgo); // Check if the cache is within the retention period
+        },
+      );
+
+      if (cachedReciters.isNotEmpty && isCacheValid) {
+        log('ReciteImpl: Returning reciters from cache (last updated: ${lastUpdatedOption.getOrElse(
+          () => DateTime.now(),
+        )})');
+        cachedReciters.sort((a, b) => a.name.compareTo(b.name));
+        return cachedReciters;
+      } else {
+        log('ReciteImpl: Cached reciters are outdated or not available');
+        await _localDataSource.clearAllReciters(); // Clear outdated cache
+      }
+
+      // If not cached or cache is outdated, fetch from the remote API
+      log('ReciteImpl: Fetching reciters from remote API');
       final reciters = await _remoteDataSource.getReciters(language: language);
       reciters.sort((a, b) => a.name.compareTo(b.name));
+
+      // Save the fetched reciters to the local cache
       await _localDataSource.saveReciters(reciters);
+
       return reciters;
     } catch (e) {
-      final reciters = await _localDataSource.getReciters();
-      reciters.sort((a, b) => a.name.compareTo(b.name));
-      return reciters;
+      // If an error occurs, try to return cached data as a fallback
+      log('ReciteImpl: Error fetching reciters: $e');
+      final cachedReciters = await _localDataSource.getReciters();
+      if (cachedReciters.isNotEmpty) {
+        log('ReciteImpl: Returning cached reciters as fallback');
+        return cachedReciters;
+      }
+
+      // If no cached data is available, rethrow the error
+      throw FetchRecitersFailedException(e.toString(), 'FETCH_RECITERS_ERROR');
     }
   }
 
