@@ -28,10 +28,10 @@ class AppDateTime {
   static final DateTime _initialRealTime = DateTime.now();
   static final DateTime _initialDebugTime = DateTime(
     _initialRealTime.year,
-    11,
-    1,
-    00,
-    -20,
+    _initialRealTime.month,
+    _initialRealTime.day-3,
+    20,
+    33+26,
     00,
   );
 
