@@ -2,20 +2,17 @@ import 'dart:async';
 import 'dart:developer';
 
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-import '../models/address_model.dart';
-
 import 'package:internet_connection_checker_plus/internet_connection_checker_plus.dart';
 
+import '../models/address_model.dart';
 import '../services/connectivity_service.dart';
 
-/// This class listens to connectivity changes and notifies its listeners.
+/// This class listens to internet‐connection changes and notifies its listeners.
 class ConnectivityProvider extends StreamNotifier<ConnectivityStatus> {
-  /// Overriding the [build] method to define how the stream of connectivity status is created.
-  /// default time for checking connection is 2 seconds.
   @override
   Stream<ConnectivityStatus> build() {
-    final internetChecker = InternetConnectionCheckerPlus.createInstance(
+    // Use the InternetConnection API (not InternetConnectionCheckerPlus)
+    final internetChecker = InternetConnection.createInstance(
       checkInterval: const Duration(seconds: 10),
     );
 
@@ -27,7 +24,8 @@ class ConnectivityProvider extends StreamNotifier<ConnectivityStatus> {
           sink.add(ConnectivityStatus.disconnected);
         },
         handleData: (status, sink) {
-          if (status == InternetConnectionStatus.connected) {
+          // Map InternetStatus to your app’s ConnectivityStatus
+          if (status == InternetStatus.connected) {
             sink.add(ConnectivityStatus.connected);
           } else {
             sink.add(ConnectivityStatus.disconnected);
@@ -37,28 +35,23 @@ class ConnectivityProvider extends StreamNotifier<ConnectivityStatus> {
     );
   }
 
-  /// [checkInternetConnection] Check the internet connection status.
-  ///
-  /// This method checks the internet connection status and updates the state accordingly.
+  /// Programmatically check internet connection once.
   Future<void> checkInternetConnection() async {
-    state = AsyncLoading();
+    state = const AsyncLoading<ConnectivityStatus>();
     state = await AsyncValue.guard(() async {
-      final internet = InternetConnectionCheckerPlus();
-      final status = await internet.hasConnection;
-      if (status) {
-        return ConnectivityStatus.connected;
-      } else {
-        return ConnectivityStatus.disconnected;
-      }
+      final internet = InternetConnection();
+      final hasInternet = await internet.hasInternetAccess;
+      return hasInternet
+          ? ConnectivityStatus.connected
+          : ConnectivityStatus.disconnected;
     });
   }
 }
 
-/// transform the stream of connectivity status to a stream of [ConnectivityStatus].
-///
-/// This provider listens to the stream of connectivity status and transforms it to a stream of [ConnectivityStatus].
-final connectivityStreamProvider = StreamProvider<ConnectivityStatus>((ref) {
-  final internetChecker = InternetConnectionCheckerPlus.createInstance(
+/// A standalone StreamProvider version, if you prefer not to use StreamNotifier.
+final connectivityStreamProvider =
+StreamProvider<ConnectivityStatus>((ref) {
+  final internetChecker = InternetConnection.createInstance(
     checkInterval: const Duration(seconds: 10),
   );
 
@@ -70,7 +63,7 @@ final connectivityStreamProvider = StreamProvider<ConnectivityStatus>((ref) {
         sink.add(ConnectivityStatus.disconnected);
       },
       handleData: (status, sink) {
-        if (status == InternetConnectionStatus.connected) {
+        if (status == InternetStatus.connected) {
           sink.add(ConnectivityStatus.connected);
         } else {
           sink.add(ConnectivityStatus.disconnected);
@@ -81,5 +74,7 @@ final connectivityStreamProvider = StreamProvider<ConnectivityStatus>((ref) {
 });
 
 /// Global provider for [ConnectivityProvider].
-/// This provider allows access to the connectivity status stream from anywhere in the app.
-final connectivityProvider = StreamNotifierProvider<ConnectivityProvider, ConnectivityStatus>(ConnectivityProvider.new);
+final connectivityProvider =
+StreamNotifierProvider<ConnectivityProvider, ConnectivityStatus>(
+  ConnectivityProvider.new,
+);
