@@ -417,18 +417,17 @@ class _SettingScreenState extends ConsumerState<SettingScreen> {
                 style: theme.textTheme.headlineSmall,
                 textAlign: TextAlign.center,
               ),
-              if (timeShiftManager.isLauncherInstalled)
-                _SettingItem(
-                  title: S.of(context).screenLock,
-                  subtitle: S.of(context).screenLockDesc,
-                  icon: Icon(Icons.power_settings_new, size: 35),
-                  onTap: () => showDialog(
-                    context: context,
-                    builder: (context) => ScreenLockModal(
-                      timeShiftManager: timeShiftManager,
-                    ),
+              _SettingItem(
+                title: S.of(context).screenLock,
+                subtitle: S.of(context).screenLockDesc,
+                icon: Icon(Icons.power_settings_new, size: 35),
+                onTap: () => showDialog(
+                  context: context,
+                  builder: (context) => ScreenLockModal(
+                    timeShiftManager: timeShiftManager,
                   ),
                 ),
+              ),
               _SettingItem(
                 title: S.of(context).appTimezone,
                 subtitle: S.of(context).descTimezone,
