@@ -7,7 +7,11 @@ import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_keyboard_visibility/flutter_keyboard_visibility.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:flutter_styled_toast/flutter_styled_toast.dart';
+import 'package:fluttertoast/fluttertoast.dart';
 import 'package:mawaqit/const/resource.dart';
+import 'package:mawaqit/src/helpers/connectivity_provider.dart';
+import 'package:mawaqit/src/models/address_model.dart';
 import 'package:mawaqit/src/pages/quran/page/schedule_screen.dart';
 import 'package:mawaqit/src/services/theme_manager.dart';
 import 'package:mawaqit/src/state_management/quran/quran/quran_notifier.dart';
@@ -27,11 +31,13 @@ import 'package:mawaqit/src/routes/routes_constant.dart';
 class AudioControlWidget extends ConsumerWidget {
   final double buttonSize;
   final double iconSize;
+  final FocusNode focusNode;
 
   const AudioControlWidget({
     Key? key,
     required this.buttonSize,
     required this.iconSize,
+    required this.focusNode,
   }) : super(key: key);
 
   bool _isWithinScheduledTime(TimeOfDay startTime, TimeOfDay endTime) {
@@ -48,6 +54,8 @@ class AudioControlWidget extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
+    final connectivityStatus = ref.watch(connectivityProvider);
+
     return ref.watch(audioControlProvider).when(
           data: (state) {
             final scheduleState = ref.watch(scheduleProvider);
@@ -55,7 +63,10 @@ class AudioControlWidget extends ConsumerWidget {
             return scheduleState.when(
               data: (schedule) {
                 final isWithinTime = _isWithinScheduledTime(schedule.startTime, schedule.endTime);
-                final shouldShow = schedule.isScheduleEnabled && isWithinTime;
+                final hasInternetConnection =
+                    connectivityStatus.hasValue && connectivityStatus.value == ConnectivityStatus.connected;
+
+                final shouldShow = schedule.isScheduleEnabled && isWithinTime && hasInternetConnection;
 
                 if (!shouldShow) return const SizedBox.shrink();
 
@@ -111,6 +122,7 @@ class _ReciterSelectionScreenState extends ConsumerState<ReciterSelectionScreen>
   late FocusScopeNode searchListFocusNode;
   late FocusNode searchFocusScopeNode;
   late FocusNode changeIntoReadingMode;
+  late FocusNode playPauseSchedule;
 
   late StreamSubscription<bool> keyboardSubscription;
 
@@ -139,6 +151,7 @@ class _ReciterSelectionScreenState extends ConsumerState<ReciterSelectionScreen>
     changeReadingModeFocusNode = FocusScopeNode(debugLabel: 'change_reading_mode_focus_scope_node');
 
     changeIntoReadingMode = FocusNode(debugLabel: 'change_into_reading_mode_focus_node');
+    playPauseSchedule = FocusNode(debugLabel: 'pla_pause_schedule');
     searchListFocusNode = FocusScopeNode(debugLabel: 'search_list_focus_scope_node');
     var keyboardVisibilityController = KeyboardVisibilityController();
 
@@ -148,7 +161,9 @@ class _ReciterSelectionScreenState extends ConsumerState<ReciterSelectionScreen>
       }
     });
     WidgetsBinding.instance.addPostFrameCallback((_) {
-      ref.read(reciteNotifierProvider.notifier);
+      if (mounted) {
+        ref.read(reciteNotifierProvider.notifier);
+      }
       _setInitialFocus();
     });
   }
@@ -212,6 +227,7 @@ class _ReciterSelectionScreenState extends ConsumerState<ReciterSelectionScreen>
     favoritesListFocusNode.onKey = _handleFavoritesListKeyEvent;
     allRecitersListFocusNode.onKey = _handleAllRecitersListKeyEvent;
     changeReadingModeFocusNode.onKey = _handleChangeReadingModeKeyEvent;
+    playPauseSchedule.onKey = _handlePlayPauseScheduleKeyEvent;
     searchListFocusNode.onKey = _handleSearchListKeyEvent;
   }
 
@@ -265,6 +281,16 @@ class _ReciterSelectionScreenState extends ConsumerState<ReciterSelectionScreen>
           return KeyEventResult.handled;
         }
         return KeyEventResult.ignored;
+      }
+    }
+    return KeyEventResult.ignored;
+  }
+
+  KeyEventResult _handlePlayPauseScheduleKeyEvent(FocusNode node, RawKeyEvent event) {
+    if (event is RawKeyDownEvent) {
+      if (event.logicalKey == LogicalKeyboardKey.arrowUp) {
+        changeIntoReadingMode.requestFocus();
+        return KeyEventResult.handled;
       }
     }
     return KeyEventResult.ignored;
@@ -468,6 +494,18 @@ class _ReciterSelectionScreenState extends ConsumerState<ReciterSelectionScreen>
     ];
   }
 
+  void showToast(String message) {
+    Fluttertoast.showToast(
+      msg: message,
+      toastLength: Toast.LENGTH_SHORT,
+      gravity: ToastGravity.BOTTOM,
+      timeInSecForIosWeb: 1,
+      backgroundColor: Colors.black87,
+      textColor: Colors.white,
+      fontSize: 16.0,
+    );
+  }
+
   Widget _buildFloatingColumn(double spacerWidth, double buttonSize, double iconSize, BuildContext context) {
     return FocusScope(
       node: changeReadingModeFocusNode,
@@ -485,6 +523,7 @@ class _ReciterSelectionScreenState extends ConsumerState<ReciterSelectionScreen>
                           autofocus: changeReadingModeFocusNode
                               .hasFocus, // it is used here because at change_reading_mode it will break due to up keybind when no result in the search
                           heroTag: 'schedule',
+                          focusColor: Theme.of(context).focusColor,
                           backgroundColor: Colors.black.withOpacity(.5),
                           child: Icon(
                             Icons.schedule,
@@ -492,10 +531,22 @@ class _ReciterSelectionScreenState extends ConsumerState<ReciterSelectionScreen>
                             size: iconSize,
                           ),
                           onPressed: () async {
-                            showDialog(
-                              context: context,
-                              builder: (BuildContext context) => ScheduleScreen(reciterList: reciter.reciters),
-                            );
+                            // Check internet connection before showing dialog
+                            await ref.read(connectivityProvider.notifier).checkInternetConnection();
+
+                            // Get the latest connectivity status after checking
+                            final updatedConnectivity = ref.read(connectivityProvider);
+                            final isConnected = updatedConnectivity.hasValue &&
+                                updatedConnectivity.value == ConnectivityStatus.connected;
+
+                            if (isConnected) {
+                              showDialog(
+                                context: context,
+                                builder: (BuildContext context) => ScheduleScreen(reciterList: reciter.reciters),
+                              );
+                            } else {
+                              showToast(S.of(context).scheduleInOnlineMode);
+                            }
                           },
                         ),
                       ),
@@ -526,6 +577,7 @@ class _ReciterSelectionScreenState extends ConsumerState<ReciterSelectionScreen>
           AudioControlWidget(
             buttonSize: buttonSize,
             iconSize: iconSize,
+            focusNode: playPauseSchedule,
           ),
         ],
       ),
