@@ -28,7 +28,10 @@ mixin NetworkConnectivity on ChangeNotifier {
   void listenToConnectivity() {
     checkIsOnline();
     _connectivityHardwareSubscription =
-        _connectivity.onConnectivityChanged.listen((ConnectivityResult result) => checkIsOnline());
+        _connectivity.onConnectivityChanged.listen((result) {
+          // Handle connectivity change
+          checkIsOnline();
+        });
 
     _connectivitySubscription = Stream.periodic(_duration).listen((event) => checkIsOnline());
   }
