@@ -1,22 +1,26 @@
 import 'dart:async';
 import 'dart:convert';
 
-import 'package:flutter/services.dart';
+import 'package:android_alarm_manager_plus/android_alarm_manager_plus.dart';
 import 'package:mawaqit/main.dart';
 import 'package:mawaqit/src/const/constants.dart';
+import 'package:mawaqit/src/domain/error/screen_on_off_exceptions.dart';
 import 'package:mawaqit/src/helpers/AppDate.dart';
 import 'package:mawaqit/src/helpers/TimeShiftManager.dart';
+import 'package:mawaqit/src/services/background_work_managers/work_manager_services.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 class TimerScheduleInfo {
   final DateTime scheduledTime;
   final String actionType; // 'screenOn' or 'screenOff'
   final bool isFajrIsha;
+  final String prayerName;
 
   TimerScheduleInfo({
     required this.scheduledTime,
     required this.actionType,
     required this.isFajrIsha,
+    required this.prayerName,
   });
 
   Map<String, dynamic> toJson() {
@@ -24,6 +28,7 @@ class TimerScheduleInfo {
       'scheduledTime': scheduledTime.toIso8601String(),
       'actionType': actionType,
       'isFajrIsha': isFajrIsha,
+      'prayerName': prayerName,
     };
   }
 
@@ -31,7 +36,8 @@ class TimerScheduleInfo {
     return TimerScheduleInfo(
       scheduledTime: DateTime.parse(json['scheduledTime']),
       actionType: json['actionType'],
-      isFajrIsha: json['isFajrIsha'],
+      isFajrIsha: json['isFajrIsha'] ?? false,
+      prayerName: json['prayerName'] ?? '',
     );
   }
 }
@@ -43,218 +49,464 @@ class ToggleScreenFeature {
 
   ToggleScreenFeature._internal();
 
-  static const String _scheduledTimersKey = TurnOnOffTvConstant.kScheduledTimersKey;
-  static final Map<String, List<Timer>> _scheduledTimers = {};
   static const String _scheduledInfoKey = 'scheduled_info_key';
-  static List<TimerScheduleInfo> _scheduleInfoList = [];
+  static const int DEFAULT_DAYS_TO_SCHEDULE = 2; // Default: Schedule for current day + 2 more days
+  static const int FAJR_ISHA_DAYS_TO_SCHEDULE = 6; // For Fajr/Isha mode: current day + 6 more days
+  static const int BACKGROUND_CHECK_ALARM_ID = 999888777; // Unique ID for background check
+  static const Duration BACKGROUND_CHECK_INTERVAL = Duration(hours: 12); // Check twice daily
 
+  /// Schedule screen toggle timers for multiple days
   static Future<void> scheduleToggleScreen(
-      bool isfajrIshaonly, List<String> timeStrings, int beforeDelayMinutes, int afterDelayMinutes) async {
-    await cancelAllScheduledTimers();
-    _scheduleInfoList.clear();
+      bool isFajrIshaOnly, List<String> timeStrings, int beforeDelayMinutes, int afterDelayMinutes) async {
+    try {
+      // Cancel any existing timers before scheduling new ones
+      await cancelAllScheduledTimers();
 
-    final now = AppDateTime.now();
-    final timeShiftManager = TimeShiftManager();
+      // Determine days to schedule based on the mode
+      final daysToSchedule = isFajrIshaOnly ? FAJR_ISHA_DAYS_TO_SCHEDULE : DEFAULT_DAYS_TO_SCHEDULE;
 
-    if (isfajrIshaonly) {
-      _scheduleForPrayer(
-        timeStrings[0],
-        now,
-        beforeDelayMinutes,
-        afterDelayMinutes,
-        true,
-        timeShiftManager,
-      );
+      // Save scheduling parameters for future rescheduling
+      await _saveSchedulingParameters(isFajrIshaOnly, timeStrings, beforeDelayMinutes, afterDelayMinutes);
 
-      _scheduleForPrayer(
-        timeStrings[5],
-        now,
-        beforeDelayMinutes,
-        afterDelayMinutes,
-        true,
-        timeShiftManager,
-      );
-    } else {
-      for (String timeString in timeStrings) {
-        _scheduleForPrayer(
-          timeString,
-          now,
-          beforeDelayMinutes,
-          afterDelayMinutes,
-          false,
-          timeShiftManager,
-        );
+      // Schedule for multiple days
+      for (int dayOffset = 0; dayOffset <= daysToSchedule; dayOffset++) {
+        await _scheduleForDay(isFajrIshaOnly, timeStrings, beforeDelayMinutes, afterDelayMinutes, dayOffset);
       }
-    }
 
-    await Future.wait([saveScheduledEventsToLocale(), toggleFeatureState(true), setLastEventDate(now)]);
+      // Schedule background check to ensure timers are still active
+      await _scheduleBackgroundCheck();
+
+      // Update feature state and last scheduled date
+      // Use a single transaction to update multiple SharedPreferences values
+      final prefs = await SharedPreferences.getInstance();
+      await Future.wait([
+        prefs.setBool(TurnOnOffTvConstant.kActivateToggleFeature, true),
+        prefs.setString(TurnOnOffTvConstant.kLastEventDate, AppDateTime.now().toIso8601String()),
+        prefs.setInt(TurnOnOffTvConstant.kMinuteBeforeKey, beforeDelayMinutes),
+        prefs.setInt(TurnOnOffTvConstant.kMinuteAfterKey, afterDelayMinutes),
+        prefs.setBool(TurnOnOffTvConstant.kIsEventsSet, true),
+      ]);
+
+      logger.i(
+          'Screen toggle scheduled successfully for ${daysToSchedule + 1} days in ${isFajrIshaOnly ? "Fajr/Isha only" : "all prayers"} mode');
+    } catch (e) {
+      logger.e('Failed to schedule toggle screen: $e');
+      throw ScheduleToggleScreenException(e.toString());
+    }
   }
 
-  static Future<void> _saveScheduleInfo() async {
+  /// Schedule timers for a specific day offset from today
+  static Future<void> _scheduleForDay(bool isFajrIshaOnly, List<String> timeStrings, int beforeDelayMinutes,
+      int afterDelayMinutes, int dayOffset) async {
+    try {
+      final List<String> prayerTimes = List.from(timeStrings);
+      prayerTimes.removeAt(1); // Remove sunrise
+      final now = AppDateTime.now();
+      final targetDate = now.add(Duration(days: dayOffset));
+      final prayerNames = ['Fajr', 'Dhuhr', 'Asr', 'Maghrib', 'Isha'];
+
+      // Create a list to collect all TimerScheduleInfo objects
+      final List<TimerScheduleInfo> schedulesToSave = [];
+
+      if (isFajrIshaOnly) {
+        // Handle Fajr (Screen ON)
+        final fajrTimeString = prayerTimes[0];
+        final fajrScheduleInfo = await _schedulePrayerAction(
+          prayerName: 'Fajr',
+          timeString: fajrTimeString,
+          targetDate: targetDate,
+          delayMinutes: beforeDelayMinutes,
+          actionType: 'screenOn',
+          isFajrIsha: true,
+          dayOffset: dayOffset,
+        );
+
+        if (fajrScheduleInfo != null) {
+          schedulesToSave.add(fajrScheduleInfo);
+        }
+
+        // Handle Isha (Screen OFF)
+        final ishaTimeString = prayerTimes[4];
+        final ishaScheduleInfo = await _schedulePrayerAction(
+          prayerName: 'Isha',
+          timeString: ishaTimeString,
+          targetDate: targetDate,
+          delayMinutes: afterDelayMinutes,
+          actionType: 'screenOff',
+          isFajrIsha: true,
+          dayOffset: dayOffset,
+        );
+
+        if (ishaScheduleInfo != null) {
+          schedulesToSave.add(ishaScheduleInfo);
+        }
+      } else {
+        // Schedule for all prayers
+        for (int i = 0; i < prayerTimes.length; i++) {
+          final prayerName = prayerNames[i];
+          final prayerTime = prayerTimes[i];
+
+          // Schedule Screen ON
+          final onScheduleInfo = await _schedulePrayerAction(
+            prayerName: prayerName,
+            timeString: prayerTime,
+            targetDate: targetDate,
+            delayMinutes: beforeDelayMinutes,
+            actionType: 'screenOn',
+            isFajrIsha: false,
+            dayOffset: dayOffset,
+          );
+
+          if (onScheduleInfo != null) {
+            schedulesToSave.add(onScheduleInfo);
+          }
+
+          // Schedule Screen OFF
+          final offScheduleInfo = await _schedulePrayerAction(
+            prayerName: prayerName,
+            timeString: prayerTime,
+            targetDate: targetDate,
+            delayMinutes: afterDelayMinutes,
+            actionType: 'screenOff',
+            isFajrIsha: false,
+            dayOffset: dayOffset,
+          );
+
+          if (offScheduleInfo != null) {
+            schedulesToSave.add(offScheduleInfo);
+          }
+        }
+      }
+
+      // Save all schedules in a single batch operation
+      if (schedulesToSave.isNotEmpty) {
+        await _saveScheduledInfoBatch(schedulesToSave);
+      }
+    } catch (e) {
+      logger.e('Failed to schedule for day with offset $dayOffset: $e');
+      throw ScheduleToggleScreenException(e.toString());
+    }
+  }
+
+  /// Schedule a single prayer action (screen on/off) and return the TimerScheduleInfo if scheduled
+  static Future<TimerScheduleInfo?> _schedulePrayerAction({
+    required String prayerName,
+    required String timeString,
+    required DateTime targetDate,
+    required int delayMinutes,
+    required String actionType,
+    required bool isFajrIsha,
+    required int dayOffset,
+  }) async {
+    try {
+      final parts = timeString.split(':');
+      final hour = int.parse(parts[0]);
+      final minute = int.parse(parts[1]);
+
+      // Create target prayer time
+      DateTime prayerDateTime = DateTime(targetDate.year, targetDate.month, targetDate.day, hour, minute);
+
+      // Apply the delay (before or after)
+      DateTime scheduledTime;
+      if (actionType == 'screenOn') {
+        scheduledTime = prayerDateTime.subtract(Duration(minutes: delayMinutes));
+      } else {
+        scheduledTime = prayerDateTime.add(Duration(minutes: delayMinutes));
+      }
+
+      final now = AppDateTime.now();
+
+      // Only schedule if the time is in the future
+      if (scheduledTime.isAfter(now)) {
+        final uniqueId = '${actionType}_${prayerName}_${dayOffset}_${DateTime.now().millisecondsSinceEpoch}';
+        final isBox = TimeShiftManager().isLauncherInstalled;
+
+        // Schedule the task
+        await WorkManagerService.registerScreenTask(uniqueId, actionType, scheduledTime.difference(now), isBox);
+
+        // Create and return the scheduled info (without saving it yet)
+        return TimerScheduleInfo(
+          scheduledTime: scheduledTime,
+          actionType: actionType,
+          isFajrIsha: isFajrIsha,
+          prayerName: prayerName,
+        );
+      }
+      return null;
+    } catch (e) {
+      logger.e('Error scheduling $actionType for $prayerName: $e');
+      throw SchedulePrayerTimeException(e.toString());
+    }
+  }
+
+  /// Save a batch of scheduled timer infos for tracking
+  static Future<void> _saveScheduledInfoBatch(List<TimerScheduleInfo> newInfoList) async {
+    try {
+      final prefs = await SharedPreferences.getInstance();
+      List<TimerScheduleInfo> existingInfoList = await _getScheduledInfoList();
+
+      // Add new infos
+      existingInfoList.addAll(newInfoList);
+
+      // Remove expired entries to keep the list manageable
+      final now = AppDateTime.now();
+      existingInfoList = existingInfoList.where((item) => item.scheduledTime.isAfter(now)).toList();
+
+      // Save back to preferences in a single operation
+      final jsonList = existingInfoList.map((item) => item.toJson()).toList();
+      await prefs.setString(_scheduledInfoKey, jsonEncode(jsonList));
+
+      // Log the number of timers saved
+      logger.i('Saved ${newInfoList.length} new timer schedules (total: ${existingInfoList.length})');
+    } catch (e) {
+      logger.e('Failed to save scheduled info batch: $e');
+    }
+  }
+
+  /// Schedule a background check to ensure all timers are still active
+  static Future<void> _scheduleBackgroundCheck() async {
+    try {
+      // Cancel any existing background check
+      await AndroidAlarmManager.cancel(BACKGROUND_CHECK_ALARM_ID);
+
+      // Schedule a new background check
+      await AndroidAlarmManager.periodic(
+        BACKGROUND_CHECK_INTERVAL,
+        BACKGROUND_CHECK_ALARM_ID,
+        backgroundCheckCallback,
+        wakeup: true,
+        rescheduleOnReboot: true,
+      );
+
+      logger.i('Background check scheduled with interval of ${BACKGROUND_CHECK_INTERVAL.inHours} hours');
+    } catch (e) {
+      logger.e('Failed to schedule background check: $e');
+    }
+  }
+
+  /// Callback for background check
+  @pragma('vm:entry-point')
+  static Future<void> backgroundCheckCallback() async {
+    try {
+      // Check if feature is active
+      if (!(await getToggleFeatureState())) {
+        logger.i('Toggle feature is inactive, skipping background check');
+        return;
+      }
+
+      // Check if rescheduling is needed using shouldReschedule
+      bool needsReschedule = await shouldReschedule();
+      if (needsReschedule) {
+        logger.i('Background check: Rescheduling needed based on shouldReschedule() criteria');
+        await _rescheduleAllTimers();
+        return;
+      }
+
+      // Additional checks if needed (existing logic)
+      final scheduledTimers = await _getScheduledInfoList();
+
+      // Check if any scheduled timers exist
+      if (scheduledTimers.isEmpty) {
+        logger.w('No scheduled timers found during background check');
+        await _rescheduleAllTimers();
+        return;
+      }
+
+      // Check if we have any future scheduled timers
+      final now = AppDateTime.now();
+      final futureTimers = scheduledTimers.where((timer) => timer.scheduledTime.isAfter(now)).toList();
+
+      if (futureTimers.isEmpty) {
+        logger.w('No future timers found during background check');
+        await _rescheduleAllTimers();
+      } else {
+        logger.i('Background check: ${futureTimers.length} future timers active');
+      }
+    } catch (e) {
+      logger.e('Background check error: $e');
+      // Try to reschedule even if there was an error
+      await _rescheduleAllTimers();
+    }
+  }
+
+  /// Reschedule all timers using saved parameters
+  static Future<void> _rescheduleAllTimers() async {
+    try {
+      logger.i('Attempting to reschedule all timers');
+
+      // Get saved scheduling parameters
+      final params = await _getSchedulingParameters();
+
+      if (params != null) {
+        // Cancel existing timers first
+        await cancelAllScheduledTimers();
+
+        // Reschedule with saved parameters
+        await scheduleToggleScreen(
+          params['isFajrIshaOnly'],
+          List<String>.from(params['timeStrings']),
+          params['beforeDelayMinutes'],
+          params['afterDelayMinutes'],
+        );
+
+        logger.i('Successfully rescheduled all timers');
+      } else {
+        logger.w('No saved scheduling parameters found for rescheduling');
+      }
+    } catch (e) {
+      logger.e('Failed to reschedule timers: $e');
+    }
+  }
+
+  /// Save scheduling parameters for future rescheduling
+  static Future<void> _saveSchedulingParameters(
+      bool isFajrIshaOnly, List<String> timeStrings, int beforeDelayMinutes, int afterDelayMinutes) async {
+    try {
+      final SharedPreferences prefs = await SharedPreferences.getInstance();
+      final Map<String, dynamic> params = {
+        'isFajrIshaOnly': isFajrIshaOnly,
+        'timeStrings': timeStrings,
+        'beforeDelayMinutes': beforeDelayMinutes,
+        'afterDelayMinutes': afterDelayMinutes,
+        'savedAt': DateTime.now().toIso8601String(),
+      };
+
+      await prefs.setString(TurnOnOffTvConstant.kScheduleParamsKey, jsonEncode(params));
+    } catch (e) {
+      logger.e('Failed to save scheduling parameters: $e');
+    }
+  }
+
+  /// Get saved scheduling parameters
+  static Future<Map<String, dynamic>?> _getSchedulingParameters() async {
+    try {
+      final SharedPreferences prefs = await SharedPreferences.getInstance();
+      final paramsString = prefs.getString(TurnOnOffTvConstant.kScheduleParamsKey);
+
+      if (paramsString != null) {
+        return jsonDecode(paramsString);
+      }
+      return null;
+    } catch (e) {
+      logger.e('Failed to get scheduling parameters: $e');
+      return null;
+    }
+  }
+
+  /// Get list of scheduled timer info
+  static Future<List<TimerScheduleInfo>> _getScheduledInfoList() async {
+    try {
+      final prefs = await SharedPreferences.getInstance();
+      final jsonString = prefs.getString(_scheduledInfoKey);
+
+      if (jsonString == null || jsonString.isEmpty) {
+        return [];
+      }
+
+      final jsonList = jsonDecode(jsonString) as List;
+      return jsonList.map((item) => TimerScheduleInfo.fromJson(item)).toList();
+    } catch (e) {
+      logger.e('Failed to get scheduled info list: $e');
+      return [];
+    }
+  }
+
+  /// Mark events as scheduled or unscheduled
+  static Future<void> _setEventsScheduled(bool isScheduled) async {
     final SharedPreferences prefs = await SharedPreferences.getInstance();
-    final scheduleData = _scheduleInfoList.map((info) => info.toJson()).toList();
-    await prefs.setString(_scheduledInfoKey, json.encode(scheduleData));
+    await prefs.setBool(TurnOnOffTvConstant.kIsEventsSet, isScheduled);
+  }
+
+  static Future<int> getBeforeDelayMinutes() async {
+    final SharedPreferences prefs = await SharedPreferences.getInstance();
+    return prefs.getInt(TurnOnOffTvConstant.kMinuteBeforeKey) ?? 0;
+  }
+
+  static Future<int> getAfterDelayMinutes() async {
+    final SharedPreferences prefs = await SharedPreferences.getInstance();
+    return prefs.getInt(TurnOnOffTvConstant.kMinuteAfterKey) ?? 0;
+  }
+
+  static Future<void> saveBeforeDelayMinutes(int minutes) async {
+    final SharedPreferences prefs = await SharedPreferences.getInstance();
+    await prefs.setInt(TurnOnOffTvConstant.kMinuteBeforeKey, minutes);
+  }
+
+  static Future<void> saveAfterDelayMinutes(int minutes) async {
+    final SharedPreferences prefs = await SharedPreferences.getInstance();
+    await prefs.setInt(TurnOnOffTvConstant.kMinuteAfterKey, minutes);
   }
 
   static Future<bool> shouldReschedule() async {
     final lastEventDate = await getLastEventDate();
     final today = AppDateTime.now();
     final isFeatureActive = await getToggleFeatureState();
-    final areEventsScheduled = await checkEventsScheduled();
+    final isEventsSet = await checkEventsScheduled();
+    final isFajrIshaOnly = await getToggleFeatureishaFajrState();
 
-    final shouldReschedule =
-        lastEventDate != null && lastEventDate.day != today.day && isFeatureActive && !areEventsScheduled;
-    return shouldReschedule;
-  }
+    // Determine the appropriate scheduling window based on the mode
+    final daysToSchedule = isFajrIshaOnly ? FAJR_ISHA_DAYS_TO_SCHEDULE : DEFAULT_DAYS_TO_SCHEDULE;
 
-  static Future<void> handleDailyRescheduling({
-    required bool isIshaFajrOnly,
-    required List<String> timeStrings,
-    required int minuteBefore,
-    required int minuteAfter,
-  }) async {
-    final shouldSchedule = await shouldReschedule();
-
-    if (shouldSchedule) {
-      await cancelAllScheduledTimers();
-      await toggleFeatureState(false);
-
-      await scheduleToggleScreen(
-        isIshaFajrOnly,
-        timeStrings,
-        minuteBefore,
-        minuteAfter,
-      );
-
-      await saveScheduledEventsToLocale();
+    // Reschedule if:
+    // 1. Feature is active but no events are scheduled
+    if (isFeatureActive && !isEventsSet) {
+      logger.i('Rescheduling needed: Feature active but no events scheduled');
+      return true;
     }
-  }
 
-  static Future<void> restoreScheduledTimers() async {
-    try {
-      final SharedPreferences prefs = await SharedPreferences.getInstance();
-      final scheduleDataString = prefs.getString(_scheduledInfoKey);
-      final isFeatureActive = await getToggleFeatureState();
+    if (lastEventDate != null && isFeatureActive) {
+      // 2. If we're approaching the end of our scheduling window
+      // Calculate days since last schedule
+      final daysSinceLastSchedule = today.difference(lastEventDate).inDays;
 
-      if (scheduleDataString == null || !isFeatureActive) {
-        return;
+      // Set appropriate reschedule threshold based on the mode
+      // For normal mode (2 days total): reschedule after 1 day
+      // For Fajr/Isha mode (6 days total): reschedule after 4-5 days
+      final rescheduleThreshold = isFajrIshaOnly ? (daysToSchedule - 2) : 1;
+
+      if (daysSinceLastSchedule >= rescheduleThreshold) {
+        logger.i(
+            'Rescheduling needed: Approaching end of scheduling window (${isFajrIshaOnly ? "Fajr/Isha mode" : "All prayers mode"})');
+        return true;
       }
 
-      final scheduleData = json.decode(scheduleDataString) as List;
-      _scheduleInfoList = scheduleData
-          .map((data) => TimerScheduleInfo.fromJson(data))
-          .where((info) => info != null) // Filter out any null entries
-          .toList();
+      // 3. Detect potential failures in scheduling system
+      // If it's been more than 24 hours since we scheduled events and
+      // we haven't detected any executed events, this could indicate a failure
+      final lastExecutedEvent = await getLastExecutedEventDate();
 
-      final now = AppDateTime.now();
-      final timeShiftManager = TimeShiftManager();
-
-      _scheduleInfoList.removeWhere((info) {
-        final isPast = info.scheduledTime.isBefore(now);
-        return isPast;
-      });
-
-      for (var info in _scheduleInfoList) {
-        try {
-          final delay = info.scheduledTime.difference(now);
-          if (delay.isNegative) {
-            continue;
-          }
-
-          final timer = Timer(delay, () {
-            try {
-              if (info.actionType == 'screenOn') {
-                timeShiftManager.isLauncherInstalled ? _toggleBoxScreenOn() : _toggleTabletScreenOn();
-              } else {
-                timeShiftManager.isLauncherInstalled ? _toggleBoxScreenOff() : _toggleTabletScreenOff();
-              }
-            } catch (e) {
-              print('Error executing timer action: $e');
-            }
-          });
-
-          final timeString = '${info.scheduledTime.hour}:${info.scheduledTime.minute}';
-          _scheduledTimers[timeString] ??= [];
-          _scheduledTimers[timeString]!.add(timer);
-        } catch (e) {
-          print('Error scheduling timer for ${info.scheduledTime}: $e');
+      if (lastExecutedEvent == null) {
+        // If no events have ever been executed but feature is active for more than a day
+        final hoursSinceLastSchedule = today.difference(lastEventDate).inHours;
+        if (hoursSinceLastSchedule > 24) {
+          logger.w('Rescheduling needed: No events have ever executed');
+          return true;
+        }
+      } else {
+        // If no events executed in the last 24 hours
+        final hoursSinceLastExecution = today.difference(lastExecutedEvent).inHours;
+        if (hoursSinceLastExecution > 24) {
+          logger.w('Rescheduling needed: No events executed in the last 24 hours');
+          return true;
         }
       }
-
-      await saveScheduledEventsToLocale();
-    } catch (e) {
-      print('Error restoring scheduled timers: $e');
     }
+
+    return false;
   }
 
-  static void _scheduleForPrayer(
-    String timeString,
-    DateTime now,
-    int beforeDelayMinutes,
-    int afterDelayMinutes,
-    bool isFajrIsha,
-    TimeShiftManager timeShiftManager,
-  ) {
-    final parts = timeString.split(':');
-    final hour = int.parse(parts[0]);
-    final minute = int.parse(parts[1]);
-
-    DateTime scheduledDateTime = DateTime(now.year, now.month, now.day, hour, minute);
-    if (scheduledDateTime.isBefore(now)) {
-      scheduledDateTime = scheduledDateTime.add(Duration(days: 1));
-    }
-
-    // Schedule screen on
-    final beforeScheduleTime = scheduledDateTime.subtract(Duration(minutes: beforeDelayMinutes));
-    if (beforeScheduleTime.isAfter(now)) {
-      _scheduleInfoList.add(TimerScheduleInfo(
-        scheduledTime: beforeScheduleTime,
-        actionType: 'screenOn',
-        isFajrIsha: isFajrIsha,
-      ));
-
-      final beforeTimer = Timer(beforeScheduleTime.difference(now), () {
-        timeShiftManager.isLauncherInstalled ? _toggleBoxScreenOn() : _toggleTabletScreenOn();
-      });
-      _scheduledTimers[timeString] = [beforeTimer];
-    } else {
-      print('Screen ON time already passed: $beforeScheduleTime');
-    }
-
-    // Schedule screen off
-    final afterScheduleTime = scheduledDateTime.add(Duration(minutes: afterDelayMinutes));
-    _scheduleInfoList.add(TimerScheduleInfo(
-      scheduledTime: afterScheduleTime,
-      actionType: 'screenOff',
-      isFajrIsha: isFajrIsha,
-    ));
-
-    final afterTimer = Timer(afterScheduleTime.difference(now), () {
-      timeShiftManager.isLauncherInstalled ? _toggleBoxScreenOff() : _toggleTabletScreenOff();
-    });
-
-    if (_scheduledTimers[timeString] != null) {
-      _scheduledTimers[timeString]!.add(afterTimer);
-    } else {
-      _scheduledTimers[timeString] = [afterTimer];
-    }
+  static Future<void> recordEventExecution() async {
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+    await prefs.setString(TurnOnOffTvConstant.kLastExecutedEventDate, DateTime.now().toIso8601String());
   }
 
-  static Future<void> loadScheduledTimersFromPrefs() async {
-    final SharedPreferences prefs = await SharedPreferences.getInstance();
-    final timersJson = prefs.getString(_scheduledTimersKey);
-
-    if (timersJson != null) {
-      final timersMap = json.decode(timersJson) as Map<String, dynamic>;
-
-      timersMap.forEach((timeString, timerDataList) {
-        _scheduledTimers[timeString] = [];
-        for (final timerData in timerDataList) {
-          final tick = timerData['tick'] as int;
-          final timer = Timer(Duration(milliseconds: tick), () {});
-          _scheduledTimers[timeString]!.add(timer);
-        }
-      });
+  static Future<DateTime?> getLastExecutedEventDate() async {
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+    final lastExecDateString = prefs.getString(TurnOnOffTvConstant.kLastExecutedEventDate);
+    if (lastExecDateString != null) {
+      return DateTime.parse(lastExecDateString);
+    } else {
+      return null;
     }
   }
 
@@ -276,86 +528,60 @@ class ToggleScreenFeature {
   }
 
   static Future<void> cancelAllScheduledTimers() async {
-    int canceledTimers = 0;
-    _scheduledTimers.forEach((timeString, timers) {
-      for (final timer in timers) {
-        timer.cancel();
-        canceledTimers++;
+    try {
+      final SharedPreferences prefs = await SharedPreferences.getInstance();
+
+      // First set the flag to false
+      await _setEventsScheduled(false);
+
+      // Get a copy of all keys to avoid concurrent modification
+      final allKeys = prefs.getKeys().toList();
+
+      // Filter keys that start with 'screen_task_id_mapping_'
+      final mappingKeys = allKeys.where((key) => key.startsWith('screen_task_id_mapping_')).toList();
+
+      // Cancel each alarm and clean up SharedPreferences
+      for (String mappingKey in mappingKeys) {
+        // Extract uniqueId from the key
+        final uniqueId = mappingKey.substring('screen_task_id_mapping_'.length);
+        // Get the alarm ID
+        final alarmIdString = prefs.getString(mappingKey);
+
+        if (alarmIdString != null) {
+          final int alarmId = int.parse(alarmIdString);
+
+          // Cancel the alarm
+          final bool success = await AndroidAlarmManager.cancel(alarmId);
+
+          if (success) {
+            logger.i('Alarm cancelled successfully: $alarmId');
+          } else {
+            logger.w('Failed to cancel alarm: $alarmId');
+          }
+
+          // Clean up SharedPreferences
+          await prefs.remove('alarm_data_$alarmId');
+          await prefs.remove(mappingKey);
+        }
       }
-    });
 
-    _scheduledTimers.clear();
-    _scheduleInfoList.clear();
+      // Also cancel the background check alarm
+      await AndroidAlarmManager.cancel(BACKGROUND_CHECK_ALARM_ID);
 
-    final SharedPreferences prefs = await SharedPreferences.getInstance();
-    await prefs.remove(_scheduledInfoKey);
-    await prefs.remove(TurnOnOffTvConstant.kScheduledTimersKey);
-    await prefs.setBool(TurnOnOffTvConstant.kIsEventsSet, false);
-  }
+      // Clear the scheduled info list
+      await prefs.remove(_scheduledInfoKey);
 
-  static Future<void> _toggleBoxScreenOn() async {
-    try {
-      await MethodChannel(TurnOnOffTvConstant.kNativeMethodsChannel)
-          .invokeMethod(TurnOnOffTvConstant.kToggleBoxScreenOn);
-    } on PlatformException catch (e) {
-      logger.e(e);
-    }
-  }
-
-  static Future<void> _toggleBoxScreenOff() async {
-    try {
-      await MethodChannel(TurnOnOffTvConstant.kNativeMethodsChannel)
-          .invokeMethod(TurnOnOffTvConstant.kToggleBoxScreenOff);
-    } on PlatformException catch (e) {
-      logger.e(e);
-    }
-  }
-
-  static Future<void> _toggleTabletScreenOn() async {
-    try {
-      await MethodChannel(TurnOnOffTvConstant.kNativeMethodsChannel)
-          .invokeMethod(TurnOnOffTvConstant.kToggleTabletScreenOn);
-    } on PlatformException catch (e) {
-      logger.e(e);
-    }
-  }
-
-  static Future<void> _toggleTabletScreenOff() async {
-    try {
-      await MethodChannel(TurnOnOffTvConstant.kNativeMethodsChannel)
-          .invokeMethod(TurnOnOffTvConstant.kToggleTabletScreenOff);
-    } on PlatformException catch (e) {
-      logger.e(e);
+      logger.i('All scheduled timers cancelled');
+    } catch (e) {
+      logger.e('Failed to cancel all scheduled timers: $e');
+      throw ScheduleToggleScreenException(e.toString());
     }
   }
 
   static Future<bool> checkEventsScheduled() async {
     final SharedPreferences prefs = await SharedPreferences.getInstance();
     final isEventsSet = prefs.getBool(TurnOnOffTvConstant.kIsEventsSet) ?? false;
-    logger.d("value$isEventsSet");
     return isEventsSet;
-  }
-
-  static Future<void> saveScheduledEventsToLocale() async {
-    final SharedPreferences prefs = await SharedPreferences.getInstance();
-
-    final scheduleData = _scheduleInfoList.map((info) => info.toJson()).toList();
-    await prefs.setString(_scheduledInfoKey, json.encode(scheduleData));
-
-    final Map<String, dynamic> timersMap = {};
-    _scheduledTimers.forEach((key, value) {
-      timersMap[key] = value.map((timer) {
-        return {
-          'isActive': timer.isActive,
-          'tick': timer.tick,
-        };
-      }).toList();
-    });
-
-    await prefs.setString(TurnOnOffTvConstant.kScheduledTimersKey, json.encode(timersMap));
-    await prefs.setBool(TurnOnOffTvConstant.kIsEventsSet, true);
-
-    logger.d("Saving into local");
   }
 
   static Future<void> setLastEventDate(DateTime date) async {
